changed namespace, downloaded colorful console package and modified help message
