+++ EchidnaEscape_01.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EchidnaEscape_01", "EchidnaEscape_01\EchidnaEscape_01.csproj", "{AC898EA3-9EEC-4B37-85D1-2D0FFFBB4A85}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{AC898EA3-9EEC-4B37-85D1-2D0FFFBB4A85}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AC898EA3-9EEC-4B37-85D1-2D0FFFBB4A85}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AC898EA3-9EEC-4B37-85D1-2D0FFFBB4A85}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AC898EA3-9EEC-4B37-85D1-2D0FFFBB4A85}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
