+++ EchidnaEscape_01/Controller/Display.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace ReferenceGame
+namespace EchidnaEscape_01
 {
 	public class Display
 	{
