+++ EchidnaEscape_01/Model/Blanket.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace ReferenceGame
+namespace EchidnaEscape_01
 {
     public class Blanket : Item
     {
