+++ EchidnaEscape_01/Model/Echidna.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace ReferenceGame
+namespace EchidnaEscape_01
 {
     public class Echidna : Item
     {
