+++ EchidnaEscape_01/Model/Game.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace ReferenceGame
+namespace EchidnaEscape_01
 {
     public enum GameState
     {
