+++ EchidnaEscape_01/Model/Honeybear.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace ReferenceGame
+namespace EchidnaEscape_01
 {
     public class Honeybear : Item
     {
