--- glisteningechidnas.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2002
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReferenceGame", "ReferenceGame\ReferenceGame.csproj", "{A492AAF5-E5D1-4829-A88E-5E99D56AA77F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A492AAF5-E5D1-4829-A88E-5E99D56AA77F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A492AAF5-E5D1-4829-A88E-5E99D56AA77F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A492AAF5-E5D1-4829-A88E-5E99D56AA77F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A492AAF5-E5D1-4829-A88E-5E99D56AA77F}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1BC91C02-AAF7-40C7-9A05-445D321A218E}
-	EndGlobalSection
-EndGlobal
