--- referencegame/ReferenceGame.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 15.00
-# Visual Studio 2017
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReferenceGame", "ReferenceGame.csproj", "{8086E905-030B-4CE5-9AEC-CBD20468E5B6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8086E905-030B-4CE5-9AEC-CBD20468E5B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8086E905-030B-4CE5-9AEC-CBD20468E5B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8086E905-030B-4CE5-9AEC-CBD20468E5B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8086E905-030B-4CE5-9AEC-CBD20468E5B6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
